(core) Extract player statistics from rakna_spelare
Move the pure counting/statistics core and exclusion-config helpers out of
the rakna_spelare.py CLI into backend/core/playerstats.py so the FastAPI
backend and the CLI share one source of truth. Terminal rendering, Colors
and argparse/main stay in rakna_spelare.py, which now imports from
core.playerstats and re-exports the moved names for backward compatibility.

Repoint the three API services (player_count_service, file_resolver,
culling_service) to import from core.playerstats. Drop the sys.path.insert
hacks in player_count_service and file_resolver that only existed to reach
rakna_spelare; culling keeps its hack for the backend-root cli_config shim.

Add core.playerstats to the PyInstaller hiddenimports. Retarget the tests'
monkeypatch paths (CONFIG_DIR/CONFIG_FILE, load_exclusion_config) to
core.playerstats where the functions now live.
