+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog is initialized from git commit history after `v1.0.0` and can be 
 ## [Unreleased]
 
 ### Changed
+- **Statistik-/räknekärnan lyft ut ur `rakna_spelare.py` till `backend/core/playerstats.py` (move-only).** Audit-steg som flyttar den delade, rena logiken (statistik + exkluderings-config) ur CLI-skriptet så FastAPI-backenden och CLI:t har en enda sanningskälla. Flyttade oförändrat till `core/playerstats.py`: `parse_filename`, `build_entries`, `segment_matches`, `bucket_counter`, `classify_deviation`, `summarize_counter`, `compute_baseline`, `compute_player_stats`, `load_exclusion_config`, `save_exclusion_config`, `resolve_exclusion_sets`, `resolve_always_markers` samt konstanterna `ALWAYS_GRUPP`/`ALWAYS_PUBLIK`/`CONFIG_DIR`/`CONFIG_FILE`. Terminalrendering (`render_bar`/`render_spark`/`print_section`), `Colors` och argparse/`main` blir kvar i `rakna_spelare.py`, som nu importerar från `core.playerstats` och re-exporterar de flyttade namnen (via `__all__`) för bakåtkompatibilitet. De tre API-tjänsterna (`player_count_service`, `file_resolver`, `culling_service`) importerar nu direkt `from core.playerstats import …`; de `sys.path.insert`-hack som bara fanns för att nå `rakna_spelare` togs bort ur `player_count_service` och `file_resolver` (culling behåller sitt hack — det behövs fortfarande för backend-rotens `cli_config`-skal). Paketering: `core.playerstats` tillagt i PyInstaller-specens `hiddenimports`. Testernas monkeypatch-mål (`CONFIG_DIR`/`CONFIG_FILE`, `load_exclusion_config`) pekar nu på `core.playerstats` där funktionerna bor.
 - **`core`-paket infört; `cli_config`/`cli_matching`/`cli_image` flyttade dit (move-only).** Audit-steg 1 av att lyfta ut delad logik ur CLI-modullagret: modulerna flyttades via `git mv` till ett nytt `backend/core/`-paket (`core/config.py`, `core/matching.py`, `core/image.py`) utan någon logik- eller funktionsnamnsändring. De gamla sökvägarna finns kvar som tunna bakåtkompatibilitetsskal (`backend/cli_config.py` m.fl. → `from core.<modul> import *`) så befintliga importörer (`hitta_ansikten.py`, `api/services/culling_service.py`, arkiverade skript) fortsätter fungera oförändrat. Paketering uppdaterad: `packages = ["core"]` i `backend/pyproject.toml` och `core.*` tillagt i PyInstaller-specens `hiddenimports`. Culling-retentiontestet pekar nu på `core.config` i stället för skalet (monkeypatch måste träffa den riktiga modulen där `load_config` läser `BASE_DIR`/`CONFIG_PATH`).
 - **Orphan CLI-skript arkiverade till `backend/scripts/archive/`.** Åtta fristående legacy-/engångsverktyg utan runtime-importörer (verifierat via grep i audit 2026-07) flyttades ut ur `backend/`-roten: `analysera_ansikten.py`, `hantera_ansikten.py`, `forfina_ansikten.py`, `ratta_ansikten.py`, `rensa_dlib.py`, `migrera_processed.py`, `inspect_encodings.py` och `update_encodings_with_filehash.py`. Inget raderades — de förblir körbara. Varje flyttat skript som importerar backend-moduler (t.ex. `faceid_db`) fick en liten `sys.path`-bootstrap i toppen som lägger till `backend/` (tre nivåer upp), så de kan köras direkt: `python scripts/archive/<verktyg>.py` från `backend/`. En `backend/scripts/archive/README.md` beskriver vad de är, varför de arkiverades och hur de körs. Följdändring: dokumentationens sökvägar till `rensa_dlib.py` m.fl. pekar nu på `scripts/archive/…`. `filer2mappar.py`, `rakna_spelare.py`, `rename_nef.py` och `hitta_ansikten.py` är fortsatt aktiva och lämnades i `backend/`-roten.
